Start cli. Add simple ALU op base. Add cli tester helper
